+++ view/frontend/web/js/widget-form.js
@@ -40,7 +40,10 @@ define([
                 });
                 form.reset();
             }).fail(function (error) {
-                var message = $.mage.__('Unexpected error');
+                var message = $.mage.__('Unexpected error.');
+                if (error.status) {
+                    message += ' ' + $.mage.__('Error Code') + ': ' + error.status;
+                }
                 if (error.responseJSON) {
                     message = error.responseJSON.message;
                 }